refactor(text-entry): Simplify answer length options
Replaces 'paragraph' and 'essay' with a single 'long' option for text entry answer length. This streamlines the UI and simplifies the underlying logic for handling longer text responses. The default answer length for new text entry questions is now 'long'.

+++ constants.tsx
@@ -63,7 +63,7 @@ export const initialSurveyData: Survey = {
           text: 'How often do you visit our café?',
           type: QuestionType.TextEntry,
           textEntrySettings: {
-            answerLength: 'short',
+            answerLength: 'long',
             placeholder: 'e.g., Once a week'
           }
         },
